+++ cli/cli/helpers/backend_for_cmd/get_backend_for_command.go
@@ -2,6 +2,7 @@ package backend_for_cmd
 
 import (
 	"github.com/kurtosis-tech/container-engine-lib/lib"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/backend_creator"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface"
 	"github.com/kurtosis-tech/stacktrace"
 )
@@ -26,7 +27,9 @@ func GetBackendForCmd(useKubernetes bool) (backend_interface.KurtosisBackend, er
 		}
 		return kubernetesBackend, nil
 	}
-	dockerBackend, err := lib.GetLocalDockerKurtosisBackend()
+	// TODO REFACTOR: we should get this backend from the config!!
+	var apiContainerModeArgs *backend_creator.APIContainerModeArgs = nil  // Not an API container
+	dockerBackend, err := backend_creator.GetLocalDockerKurtosisBackend(apiContainerModeArgs)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "Expected to be able to get Docker backend, instead a non-nil error was returned")
 	}