+++ cli/docs/changelog.md
@@ -2,6 +2,10 @@
 ### Changes
 * Upgraded to container-engine-lib 0.21.0, core 1.47.0, and engine 1.19.0 which use the new method of registering services
 
+### Breaking Changes
+* Upgraded the engine to version 1.19.0
+    * Users should run `kurtosis engine restart` after upgrading
+
 # 0.16.2
 ### Features
 * Can use Kurtosis Core YAML config file format V1 to specify metrics preferences